feat: Implement item management functionality
- Added Item entity with fields for seller, category, brand, name, description, price, stock, condition, shipping details, status, likes count, and view count.
- Created ItemImage entity to handle item images with display order.
- Developed ItemRepository interface for CRUD operations on items.
- Implemented MySQL-based ItemRepository with methods for creating, finding, updating, deleting, and incrementing view counts of items.
- Introduced ItemUseCase for business logic related to items, including creation, retrieval, updating, deletion, and purchasing.
- Added validation methods for item properties (name, price, description).
- Created mock repository for unit testing ItemUseCase.
- Implemented unit tests for item creation, retrieval, updating, deletion, and purchasing scenarios.
- Added response structures for item and item images to facilitate API responses.

+++ migrations/000006_create_wallet_notifications.up.sql
@@ -4,8 +4,8 @@ CREATE TABLE wallet_transactions (
     user_id BIGINT NOT NULL,
     related_id BIGINT NULL,
     type ENUM('sales_deposit', 'purchase', 'withdrawal', 'point_grant', 'point_expire', 'refund') NOT NULL,
-    amount DECIMAL(10, 2) NOT NULL,
-    balance_snapshot DECIMAL(10, 2) NOT NULL,
+    amount BIGINT NOT NULL,
+    balance_snapshot BIGINT NOT NULL,
     description VARCHAR(255),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
